requirements.txt
attrs==17.3.0
certifi==2017.11.5
chardet==3.0.4
cytoolz==0.9.0
decorator==4.1.2
eth-abi==0.5.0
eth-keyfile==0.4.1
eth-keys==0.1.0b4
eth-tester==0.1.0b9
eth-utils==0.7.3
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
jupyter-client==5.2.0
jupyter-core==4.4.0
numpy==1.13.3
pandas==0.21.1
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
-e git+https://github.com/corpetty/py-etherscan-api.git@a783f720a6175b227cdfe8dc3040ebad3022d79f#egg=py_etherscan_api
py-solc==2.1.0
pycryptodome==3.4.7
Pygments==2.2.0
pylru==1.0.9
pysha3==1.0.2
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
simplegeneric==0.8.1
six==1.11.0
toolz==0.9.0
tornado==4.5.2
tqdm==4.19.5
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
web3==3.16.4
